+++ react/SearchResult.js
@@ -1,37 +1,31 @@
-import PropTypes from 'prop-types'
+import './global.css'
+
 import React, { Component } from 'react'
 import { ProductSummary } from 'vtex.product-summary'
 
 import SearchResultInfiniteScroll from './components/SearchResultInfiniteScroll'
-import { queryShape, schemaPropsTypes } from './constants/propTypes'
-import { createMap, reversePagesPath } from './constants/SearchHelpers'
+import { searchResultPropTypes } from './constants/propTypes'
+import SortOptions from './constants/SortOptions'
 
-const DEFAULT_PAGE = 1
 const DEFAULT_MAX_ITEMS_PER_LINE = 5
 const DEFAULT_MAX_ITEMS_PER_PAGE = 10
 
+/**
+ * Search Result Component.
+ */
 export default class SearchResult extends Component {
-  static propTypes = {
-    params: PropTypes.shape({
-      /** Search's term, e.g: eletronics. */
-      term: PropTypes.string,
-      /** Department param. */
-      department: PropTypes.string,
-      /** Category param. */
-      category: PropTypes.string,
-      /** Subcategory param. */
-      subcategory: PropTypes.string,
-    }),
-    query: queryShape,
-    /** Internal route path. e.g: 'store/search' */
-    pagesPath: PropTypes.string,
-    ...schemaPropsTypes,
+  constructor(props) {
+    super(props)
+    const { maxItemsPerPage = DEFAULT_MAX_ITEMS_PER_PAGE } = props
+    props.setContextVariables({ maxItemsPerPage })
   }
 
+  static propTypes = searchResultPropTypes
+
   static defaultProps = {
-    maxItemsPerLine: DEFAULT_MAX_ITEMS_PER_LINE,
+    orderBy: SortOptions[0].value,
+    rest: '',
     maxItemsPerPage: DEFAULT_MAX_ITEMS_PER_PAGE,
-    pagesPath: 'store/search',
   }
 
   static uiSchema = {
@@ -70,32 +64,6 @@ export default class SearchResult extends Component {
   }
 
   render() {
-    const {
-      maxItemsPerLine,
-      maxItemsPerPage,
-      summary,
-      pagesPath,
-      query: {
-        order: orderBy,
-        page: pageProps,
-        map: mapProps,
-        rest,
-      },
-    } = this.props
-    const pathName = reversePagesPath(pagesPath, this.props.params)
-    const map = mapProps || createMap(pathName, rest)
-    const page = (pageProps ? parseInt(pageProps) : DEFAULT_PAGE)
-    const containerProps = {
-      path: pathName,
-      map,
-      rest,
-      page,
-      orderBy,
-      maxItemsPerLine,
-      maxItemsPerPage,
-      summary,
-      pagesPath,
-    }
-    return <SearchResultInfiniteScroll {...containerProps} />
+    return <SearchResultInfiniteScroll {...this.props} />
   }
 }