+++ react/components/SearchResultInfiniteScroll.js
@@ -1,17 +1,10 @@
-import '../global.css'
-
-import PropTypes from 'prop-types'
 import React, { Component } from 'react'
-import { compose, graphql } from 'react-apollo'
 import InfiniteScroll from 'react-infinite-scroll-component'
 import { Spinner } from 'vtex.styleguide'
 
 import VTEXClasses from '../constants/CSSClasses'
-import { facetsQueryShape, mapType, orderType, schemaPropsTypes, searchQueryShape } from '../constants/propTypes'
-import { getPagesArgs, stripPath, joinPathWithRest, getCategoriesFromQuery, findInTree } from '../constants/SearchHelpers'
-import SortOptions from '../constants/SortOptions'
-import facetsQuery from '../graphql/facetsQuery.gql'
-import searchQuery from '../graphql/searchQuery.gql'
+import { searchResultPropTypes } from '../constants/propTypes'
+import { findInTree, getCategoriesFromQuery, getPagesArgs, stripPath } from '../constants/SearchHelpers'
 import Gallery from './Gallery'
 import SearchFilter from './SearchFilter'
 import SearchHeader from './SearchHeader'
@@ -35,10 +28,14 @@ const MAP_SEPARATOR = ','
 /**
  * Search Result Component.
  */
-class SearchResultInfiniteScroll extends Component {
+export default class SearchResultInfiniteScroll extends Component {
+  static propTypes = searchResultPropTypes
+
   getLinkProps = ({ opt, type, isSelected, ordenation, pageNumber }) => {
     const { path, rest, map, pagesPath } = this.props
-    let { variables: { orderBy } } = this.props.searchQuery
+    let {
+      variables: { orderBy },
+    } = this.props.searchQuery
     orderBy = ordenation || orderBy
     return getPagesArgs(
       { name: opt && opt.Name, type, link: opt && opt.Link },
@@ -53,7 +50,9 @@ class SearchResultInfiniteScroll extends Component {
   getCategories() {
     const {
       facetsQuery,
-      searchQuery: { variables: { query, map } },
+      searchQuery: {
+        variables: { query, map },
+      },
       pagesPath,
     } = this.props
     const { CategoriesTrees: tree } = facetsQuery.facets
@@ -70,7 +69,9 @@ class SearchResultInfiniteScroll extends Component {
   renderSearchFilters() {
     if (!this.props.facetsQuery || !this.props.facetsQuery.facets) return
 
-    const { facetsQuery: { facets: facetsProps } } = this.props
+    const {
+      facetsQuery: { facets: facetsProps },
+    } = this.props
     const facets = { ...facetsProps }
     const selecteds = this.getSelecteds()
     delete facets[FACETS_KEYS.Departments]
@@ -229,7 +230,10 @@ class SearchResultInfiniteScroll extends Component {
         hasMore={products.length < recordsFiltered}>
         <div className={`${VTEXClasses.MAIN_CLASS} w-100 pa3 dib`}>
           <div className="w-100 w-30-m w-20-l fl pa3">
-            <SelectedFilters selecteds={selecteds} getLinkProps={this.getLinkProps} />
+            <SelectedFilters
+              selecteds={selecteds}
+              getLinkProps={this.getLinkProps}
+            />
             {this.renderSearchFilters()}
           </div>
           <div className="w-100 w-70-m w-80-l fl">
@@ -249,58 +253,3 @@ class SearchResultInfiniteScroll extends Component {
     )
   }
 }
-
-const SearchResultInfiniteScrollWithData = compose(
-  graphql(facetsQuery, {
-    name: 'facetsQuery',
-    options: (props) => {
-      const { path, rest, map } = props
-      const query = joinPathWithRest(path, rest)
-      const facets = `${query}?map=${map}`
-      return ({
-        variables: { facets },
-      })
-    },
-  }),
-  graphql(searchQuery, {
-    name: 'searchQuery',
-    options: props => {
-      const { path, rest, map } = props
-      const query = joinPathWithRest(path, rest)
-      const orderBy = props.orderBy
-      const from = (props.page - 1) * props.maxItemsPerPage
-      const to = from + props.maxItemsPerPage - 1
-      return {
-        variables: { query, map, orderBy, from, to },
-        notifyOnNetworkStatusChange: true,
-      }
-    },
-  })
-)(SearchResultInfiniteScroll)
-
-SearchResultInfiniteScroll.propTypes = SearchResultInfiniteScrollWithData.propTypes = {
-  /** Internal route path. e.g: 'store/search' */
-  pagesPath: PropTypes.string,
-  /** Path param. e.g: eletronics/smartphones */
-  path: PropTypes.string,
-  /** Map param. e.g: c,c */
-  map: mapType.isRequired,
-  /** Rest param. e.g: Android,Samsung */
-  rest: mapType.isRequired,
-  /** Search result page. */
-  page: PropTypes.number.isRequired,
-  /** Search result ordernation. */
-  orderBy: orderType,
-  /** Facets graphql query. */
-  facetsQuery: facetsQueryShape,
-  /** Search graphql query. */
-  searchQuery: searchQueryShape,
-  ...schemaPropsTypes,
-}
-
-SearchResultInfiniteScroll.defaultProps = SearchResultInfiniteScrollWithData.defaultProps = {
-  orderBy: SortOptions[0].value,
-  rest: '',
-}
-
-export default SearchResultInfiniteScrollWithData